+++ carp_mobile_sensing/example/lib/paper_example.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 
 void sensing() async {
   // create the study
-  Study study = Study('DF#4dD-1', 'user@gmail.com',
+  Study study = Study("2", 'user@gmail.com',
       name: 'A simple example study',
       dataEndPoint: FileDataEndPoint()
         ..bufferSize = 500 * 1000