+++ carp_webservices/lib/carp_service/carp_datapoint.dart
@@ -14,6 +14,9 @@ part of carp_services;
 class CARPDataPoint {
   /// A unique, server-side generated ID for this data point.
   /// [null] if this data point is not yet stored in the CARP server.
+  ///
+  /// TODO - note that on the CARP web service implementation this is an int -- but should be a string.
+  /// See issue #16 >> https://github.com/cph-cachet/carp.webservices/issues/16
   int id;
 
   /// The unique study id that this data point belongs to.
@@ -37,7 +40,7 @@ class CARPDataPoint {
 
   /// Create a [CARPDataPoint] based on a [DataPoint] generated in the CARP Mobile Sensing Framework.
   CARPDataPoint.fromDataPoint(DataPoint dataPoint) {
-    CARPDataPointHeader header = new CARPDataPointHeader(dataPoint.header.studyId, dataPoint.header.userId);
+    CARPDataPointHeader header = new CARPDataPointHeader(dataPoint.header.studyId.toString(), dataPoint.header.userId);
     header.startTime =
         (dataPoint.body is CARPDatum) ? (dataPoint.body as CARPDatum).timestamp.toUtc() : new DateTime.now().toUtc();
     header.dataFormat = dataPoint.header.dataFormat;
@@ -47,8 +50,11 @@ class CARPDataPoint {
   }
 
   /// Create a [CARPDataPoint] based on a [CARPDatum] generated in the CARP Mobile Sensing Framework.
+  ///
+  /// TODO - once the restriction on only integers as study id on the server side is fixed, remove the assert below.
   CARPDataPoint.fromDatum(String studyId, String userId, CARPDatum datum) {
-    CARPDataPointHeader header = new CARPDataPointHeader(studyId, userId);
+    assert(int.tryParse(studyId) != null, 'Study ID ($studyId) for the CARP Web Services can only be an integer.');
+    CARPDataPointHeader header = new CARPDataPointHeader(studyId.toString(), userId);
     header.startTime = (datum is CARPDatum) ? datum.timestamp.toUtc() : new DateTime.now().toUtc();
     header.dataFormat = datum.format;
 
@@ -89,7 +95,7 @@ class CARPDataPointHeader {
   /// The CARP data format. See [DataFormat] and [NameSpace].
   DataFormat dataFormat;
 
-  // Create a new [CARPDataPointHeader]. [studyId] and [userId] are required.
+  /// Create a new [CARPDataPointHeader]. [studyId] and [userId] are required.
   CARPDataPointHeader(this.studyId, this.userId, {this.deviceRoleName, this.triggerId, this.startTime, this.endTime}) {
     if (startTime != null) startTime.toUtc();
     if (endTime != null) endTime.toUtc();