+++ carp_webservices/lib/carp_service/carp_service.g.dart
@@ -8,11 +8,12 @@ part of carp_services;
 
 CARPDataPoint _$CARPDataPointFromJson(Map<String, dynamic> json) {
   return CARPDataPoint(
-      json['carp_header'] == null
-          ? null
-          : CARPDataPointHeader.fromJson(
-              json['carp_header'] as Map<String, dynamic>),
-      json['carp_body'] as Map<String, dynamic>)
+    json['carp_header'] == null
+        ? null
+        : CARPDataPointHeader.fromJson(
+            json['carp_header'] as Map<String, dynamic>),
+    json['carp_body'] as Map<String, dynamic>,
+  )
     ..id = json['id'] as int
     ..studyId = json['study_id'] as int
     ..createdByUserId = json['created_by_user_id'] as int;
@@ -37,15 +38,17 @@ Map<String, dynamic> _$CARPDataPointToJson(CARPDataPoint instance) {
 
 CARPDataPointHeader _$CARPDataPointHeaderFromJson(Map<String, dynamic> json) {
   return CARPDataPointHeader(
-      json['study_id'] as String, json['user_id'] as String,
-      deviceRoleName: json['device_role_name'] as String,
-      triggerId: json['trigger_id'] as String,
-      startTime: json['start_time'] == null
-          ? null
-          : DateTime.parse(json['start_time'] as String),
-      endTime: json['end_time'] == null
-          ? null
-          : DateTime.parse(json['end_time'] as String))
+    json['study_id'] as String,
+    json['user_id'] as String,
+    deviceRoleName: json['device_role_name'] as String,
+    triggerId: json['trigger_id'] as String,
+    startTime: json['start_time'] == null
+        ? null
+        : DateTime.parse(json['start_time'] as String),
+    endTime: json['end_time'] == null
+        ? null
+        : DateTime.parse(json['end_time'] as String),
+  )
     ..uploadTime = json['upload_time'] == null
         ? null
         : DateTime.parse(json['upload_time'] as String)